probe_client/requirements.txt
websockets>=16.0,<17.0
PyYAML>=6.0,<7.0
psutil>=6.0,<7.0
mss>=10.0,<11.0
Pillow>=11.0,<12.0
